1.优化服务端：创建CCommand命令类 2.优化服务端：创建CTool工具类
