+++ RemoteCtrl/RemoteCtrl.sln
@@ -9,7 +9,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RemoteClient", "RemoteClien
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "解决方案项", "解决方案项", "{058BD256-1861-44A1-A687-33A638403F5B}"
 	ProjectSection(SolutionItems) = preProject
-		远程控制系统设计.drawio = 远程控制系统设计.drawio
+		远程控制系统设计.mdj = 远程控制系统设计.mdj
 	EndProjectSection
 EndProject
 Global
