+++ RemoteCtrl/RemoteCtrl/Tool.cpp
@@ -1,0 +1,15 @@
+#include "pch.h"
+#include "Tool.h"
+
+void CTool::Dump(BYTE* pData, size_t nSize) {
+    std::string strOut;
+    for (size_t i = 0; i < nSize; i++) {
+        char buf[8] = "";
+        if (i % 15 == 0) strOut += "\n";
+        snprintf(buf, sizeof(buf), "%02X ", pData[i] & 0xFF);
+        strOut += buf;
+    }
+    strOut += "\n";
+    OutputDebugStringA(strOut.c_str());
+}
+
