+++ RemoteCtrl/RemoteCtrl/Tool.h
@@ -1,0 +1,8 @@
+#pragma once
+class CTool
+{
+public:
+	static void Dump(BYTE* pData, size_t nSize);
+
+};
+
